Restructure async route code
